refactor(core,dom): split Navigator into context-agnostic base + DOMNavigator
Model the navigator the way the context abstraction works: the base `Navigator`
(moved into core's `core/` folder) now owns only the view model — the
`{ k, x, y }` transform, brush, and the imperative pan/zoom/frame commands — and
attaches no DOM listeners of its own, so it can back non-DOM environments or be
driven programmatically.

The DOM listener wiring moves to a `DOMNavigator` subclass in `@ripl/dom`, which
translates real wheel/pointer/touch gestures into the base commands and exports
the `createNavigator` factory. The gesture model is Figma-like: click-and-hold
(left or middle, with or without ⌘/Ctrl) pans, wheel/pinch zooms toward the
cursor, and shift-drag brushes. Because translation is unbounded and the pointer
is captured during a drag, the view can roam to content outside the viewport,
then re-frame with the new `centerOn`/`fitBounds` helpers.
