--- packages/core/src/interaction/index.ts
@@ -1,1 +1,0 @@
-export * from './navigator';
