+++ packages/dom/src/index.ts
@@ -1,4 +1,5 @@
 export * from './dom';
 export * from './context';
 export * from './export';
+export * from './navigator';
 export * from './vdom';
